+++ fsspec/__init__.py
@@ -5,6 +5,7 @@ from .registry import get_filesystem_class, registry, filesystem
 from .mapping import FSMap, get_mapper
 from .core import open_files, get_fs_token_paths, open
 from . import caching
+from .dircache import DirCache, StaleKeyError
 
 __version__ = get_versions()["version"]
 del get_versions
@@ -13,6 +14,8 @@ del get_versions
 __all__ = [
     "AbstractFileSystem",
     "FSMap",
+    "DirCache",
+    "StaleKeyError",
     "filesystem",
     "get_filesystem_class",
     "get_fs_token_paths",